refactor(typescript): convert tags/controller
